LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-removetrack.https-expected.txt

When prompted, accept to share your audio stream, then your video stream.

Description

This test checks that removinging a track from a MediaStream works as expected.



PASS Tests that a removal from a MediaStream works as expected
PASS Test that removal from a MediaStream fires ended on media elements (video first)
FAIL Test that removal from a MediaStream fires ended on media elements (audio first) assert_equals: audio element ended because no more audio tracks expected true but got false
